swapped components
